requirements.txt
camelot-py==1.0.9
opencv-python-headless==4.12.0.88
pdfminer.six==20250506
pdfplumber==0.11.7
PyPDF2==3.0.1
pillow==11.3.0
openpyxl==3.1.5

PyQt6==6.9.1
PyQt6-Qt6==6.9.2
PyQt6_sip==13.10.2

numpy==2.2.6
pandas==2.3.3
matplotlib==3.10.6
python-dateutil==2.9.0.post0

streamlit==1.50.0
